+++ libraries/e-signature/fot.config.ts
@@ -10,5 +10,5 @@ export default defineFotConfig({
       "@f-o-t/digital-certificate",
       "react",
    ],
-   plugins: ["react", "worker", "worker-entry"],
+   plugins: ["react", "worker"],
 });